+++ src/act.off.c
@@ -880,8 +880,8 @@ void do_bash(struct char_data *ch, char *argument, int cmd)
   percent=number(1,101); /* 101% is a complete failure */
   
   /* some modifications to account for dexterity, and level */
-  percent -= dex_app[GET_DEX(ch)].reaction*10;
-  percent += dex_app[GET_DEX(victim)].reaction*10;
+  percent -= dex_app[(int)GET_DEX(ch)].reaction*10;
+  percent += dex_app[(int)GET_DEX(victim)].reaction*10;
   if (GetMaxLevel(victim) > 12) {
     percent += ((GetMaxLevel(victim)-10) * 5);
   }
